Add some tests for the simple individual.
This required some internal functions to be exported, and I'd rather
they didn't need to be, but the methods on the instance take an RNG
and it would be complicated to give each method an RNG that returned
the generated value. Another option would be making the RNG instance a
property of the data type.