Implement subtree mutation, add Divide back.
The subtree mutation replaces a random node in the tree with a random
tree of the same maximum size as the initial random trees. This finds
a solution to x^2 correctly.

I've set the result of 0/x x/0 and 0/0 to a large value to avoid
it. But trees containing this subtree should probably be given a low
fitness instead.